Slightly optimize CPU switch

+++ core/src/tf/veriny/kl27/cpu/CPU.kt
@@ -63,11 +63,11 @@ class CPU(f: K27File) {
         // MAIN INTERPRETER BLOCK
         // This runs the actual code.
         //println(instruction.opcode.toString())
-        when (instruction.opcode) {
-            0x0.toShort() -> {
+        when (instruction.opcode.toInt()) {
+            0x0 -> {
                 // no-op, do nothing
             }
-            0x1.toShort() -> {
+            0x1 -> {
                 // JMPL, jump to label
                 val offset = this.memory.getLabelOffset(instruction.opcode)
                 // we need to set it to 0x01000 + offset
